import/requirements.txt
PyYAML==6.0.3
requests==2.33.0
python-frontmatter==1.1.0
isodate==0.7.2